+++ web/gui/version.txt
@@ -1,1 +1,1 @@
-252b1472708ea159f00037a1c39cd00d2ac10702
+12641f27666d468dd160dd87746bbb1bb1c5ecd5
